requirements.txt
loguru
tencentcloud-sdk-python
opencv-python